requirements.txt
appdirs==1.4.3
ebaysdk==2.1.0
Flask==0.10.1
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.0
MarkupSafe==0.23
packaging==16.8
pylibmc==1.5.2
pyparsing==2.2.0
PyYAML==3.11
requests==2.4.3
six==1.10.0
Werkzeug==0.9.6
